마트 과제

+++ article/src/com/ktdsuniversity/edu/mart/Customer.java
@@ -30,7 +30,7 @@ public class Customer {
 		this.money = money;
 	}
 	
-	public double getPoint() {
+	public int getPoint() {
 		return point;
 	}
